LayoutTests/fast/dom/NodeList/node-list-array-class.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<script>

description('This tests that NodeLists have Array in the prototype chain');

var nodeList = document.childNodes;
shouldBe('nodeList.__proto__', 'NodeList.prototype');
shouldBe('nodeList.__proto__.__proto__', 'Array.prototype');

shouldBe('NodeList.prototype.__proto__', 'Array.prototype');

var myProto = new String('abc');
NodeList.prototype.__proto__ = myProto;
shouldBe('NodeList.prototype.__proto__', 'myProto');

var myPrototype = {};
NodeList.prototype = myPrototype;
shouldNotBe('NodeList.prototype', 'myPrototype');

</script>
<script src="../../js/resources/js-test-post.js"></script>
