LayoutTests/fast/dom/NodeList/node-list-array-class-expected.txt

This tests that NodeLists have Array in the prototype chain

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS nodeList.__proto__ is NodeList.prototype
PASS nodeList.__proto__.__proto__ is Array.prototype
PASS NodeList.prototype.__proto__ is Array.prototype
PASS NodeList.prototype.__proto__ is myProto
PASS NodeList.prototype is not myPrototype
PASS successfullyParsed is true

TEST COMPLETE
